backend/requirements.txt
fastapi
uvicorn
google-genai
pypdf
python-multipart
python-dotenv
chromadb
